Modules Users

+++ src/main/java/vn/kyler/job_hunter/service/UserService.java
@@ -2,6 +2,7 @@ package vn.kyler.job_hunter.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,9 +10,11 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import vn.kyler.job_hunter.domain.User;
-import vn.kyler.job_hunter.domain.dto.Meta;
-import vn.kyler.job_hunter.domain.dto.ResultPaginationDTO;
+import vn.kyler.job_hunter.domain.response.ResUserDTO;
+import vn.kyler.job_hunter.domain.response.ResultPaginationDTO;
 import vn.kyler.job_hunter.repository.UserRepository;
+import vn.kyler.job_hunter.service.exception.EmailExistsException;
+import vn.kyler.job_hunter.service.exception.NotFoundException;
 
 @Service
 public class UserService {
@@ -22,17 +25,27 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public User handleCreateUser(User user) {
+    public User handleCreateUser(User user) throws EmailExistsException {
+        boolean existsEmail = this.userRepository.existsByEmail(user.getEmail());
+        if (existsEmail) {
+            throw new EmailExistsException("Email " + user.getEmail() + " already exists");
+        }
         return this.userRepository.save(user);
     }
 
-    public void handleDeleteUser(long id) {
+    public void handleDeleteUser(long id) throws NotFoundException {
+        if (!this.userRepository.existsById(id)) {
+            throw new NotFoundException("User with id " + id + " not found");
+        }
         this.userRepository.deleteById(id);
     }
 
-    public User handleGetUser(long id) {
+    public User handleGetUser(long id) throws NotFoundException {
         Optional<User> user = this.userRepository.findById(id);
-        return user.isPresent() ? user.get() : null;
+        if (!user.isPresent()) {
+            throw new NotFoundException("User with id " + id + " not found");
+        }
+        return user.get();
     }
 
     public User handleGetUserByEmail(String email) {
@@ -46,8 +59,24 @@ public class UserService {
 
     public ResultPaginationDTO handleGetAllUser(Specification<User> specification, Pageable pageable) {
         Page<User> userPage = this.userRepository.findAll(specification, pageable);
+        List<ResUserDTO> userDTOs = userPage.getContent().stream()
+                .map(user -> {
+                    ResUserDTO userDTO = new ResUserDTO();
+                    userDTO.setId(user.getId());
+                    userDTO.setName(user.getName());
+                    userDTO.setEmail(user.getEmail());
+                    userDTO.setAge(user.getAge());
+                    userDTO.setGender(user.getGender());
+                    userDTO.setAddress(user.getAddress());
+                    userDTO.setCreatedAt(user.getCreatedAt());
+                    userDTO.setUpdatedAt(user.getUpdatedAt());
+                    userDTO.setCreatedBy(user.getCreatedBy());
+                    userDTO.setUpdatedBy(user.getUpdatedBy());
+                    return userDTO;
+                })
+                .collect(Collectors.toList());
         ResultPaginationDTO resultPaginationDTO = new ResultPaginationDTO();
-        Meta meta = new Meta();
+        ResultPaginationDTO.Meta meta = new ResultPaginationDTO.Meta();
 
         meta.setPageNumber(userPage.getNumber() + 1);
         meta.setPageSize(userPage.getSize());
@@ -55,20 +84,54 @@ public class UserService {
         meta.setTotalElements(userPage.getTotalElements());
 
         resultPaginationDTO.setMeta(meta);
-        resultPaginationDTO.setResult(userPage.getContent());
+        resultPaginationDTO.setResult(userDTOs);
         return resultPaginationDTO;
     }
 
-    public User handleUpdateUser(User user) {
+    public User handleUpdateUser(User user) throws NotFoundException {
         Optional<User> userOptional = this.userRepository.findById(user.getId());
         if (userOptional.isPresent()) {
             User userToUpdate = userOptional.get();
             userToUpdate.setName(user.getName());
-            userToUpdate.setEmail(user.getEmail());
-            userToUpdate.setPassword(user.getPassword());
+            // userToUpdate.setEmail(user.getEmail());
+            userToUpdate.setAge(user.getAge());
+            userToUpdate.setGender(user.getGender());
+            userToUpdate.setAddress(user.getAddress());
+
             return this.userRepository.save(userToUpdate);
         } else {
+            throw new NotFoundException("User with id " + user.getId() + " not found");
+        }
+    }
+
+    public ResUserDTO handleConvertToUserDTO(User user) {
+        ResUserDTO userDTO = new ResUserDTO();
+        userDTO.setId(user.getId());
+        userDTO.setName(user.getName());
+        userDTO.setEmail(user.getEmail());
+        userDTO.setAge(user.getAge());
+        userDTO.setGender(user.getGender());
+        userDTO.setAddress(user.getAddress());
+        userDTO.setCreatedAt(user.getCreatedAt());
+        userDTO.setUpdatedAt(user.getUpdatedAt());
+        userDTO.setCreatedBy(user.getCreatedBy());
+        userDTO.setUpdatedBy(user.getUpdatedBy());
+        return userDTO;
+    }
+
+    public void handleUpdateUserToken(String email, String token) {
+        User user = this.handleGetUserByEmail(email);
+        if (user != null) {
+            user.setRefreshToken(token);
+            this.userRepository.save(user);
+        }
+    }
+
+    public User handleGetUserByRefreshTokenAndEmail(String refreshToken, String email) {
+        Optional<User> userOptional = this.userRepository.findByRefreshTokenAndEmail(refreshToken, email);
+        if (!userOptional.isPresent()) {
             return null;
         }
+        return userOptional.get();
     }
 }
